requirements.txt
uvicorn==0.34.0
fastapi==0.115.6
python-dotenv==1.0.1
boto3==1.35.87
httpx==0.28.1
Jinja2==3.1.5
python-multipart==0.0.20
